+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: ignore folder feature added
 
 1.1.0 (2022-03-14)
 ------------------